% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{find_population_weights}
\alias{find_population_weights}
\title{Calculate population weights}
\usage{
find_population_weights(df_experiment, df_popweights)
}
\arguments{
\item{df_experiment}{Original dataset}

\item{df_popweights}{Dataset of population weights}
}
\value{
A dataset of weights for each age_category X region X gender?
}
\description{
Calculate population weights
}
