% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmtalent_df}
\alias{mmtalent_df}
\title{Data from mmtalent, an online experiment
A dataset from the online experiment used in the paper
"Talent and luck - An experimental study on inequality
acceptance" by Björn Bartling, Alexander W. Cappelen,
Ingvild L. Skarpeid, Erik Ø. Sørensen and Bertil Tungodden.}
\format{}
\source{
Based on experiment by the authors.
}
\usage{
mmtalent_df
}
\description{
Data from mmtalent, an online experiment
A dataset from the online experiment used in the paper
"Talent and luck - An experimental study on inequality
acceptance" by Björn Bartling, Alexander W. Cappelen,
Ingvild L. Skarpeid, Erik Ø. Sørensen and Bertil Tungodden.
}
\keyword{datasets}
