% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{populationweights2016}
\alias{populationweights2016}
\title{Census data on population in Agegroup x region x gender cells}
\format{A data frame with 40 observations and 4 variables:
\describe{
\item{region}{The four census regions}
\item{gender}{"male" or "female", characters}
\item{age_category}{ (characters) \itemize{
 \item [18,35)
 \item [35,45),
 \item [45,55),
 \item [55,65),
 \item [65, 120)
 }}
\item{n2016}{Population in 2016 (numeric)}
}}
\source{
Based on the sc-est2016-alldata6.csv file from https://www2.census.gov/programs-surveys/popest/datasets/2010-2016/state/asrh/ downloaded November 28, 2017.
}
\usage{
populationweights2016
}
\description{
Census data on population in Agegroup x region x gender cells
}
\keyword{datasets}
