% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{prepare_data}
\alias{prepare_data}
\title{Do basic data transforms}
\usage{
prepare_data(df, df_popweights)
}
\arguments{
\item{df}{Original dataset}

\item{df_popweights}{Dataset of population weights per age_category X region X gender}
}
\value{
Transformed dataset
}
\description{
Do basic data transforms
}
